[Rename] 파일 옮김
- post-place.tsx 파일을 src/app/post-place로 옮기고 파일명 page.tsx로 변경
