+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Nav from '@/components/nav-items';
-import PostPlace from '@/features/post-place/components/post-place'
-import SearchPage from '@/features/search-page/components/search-page';
+import PostPlace from '@/app/post-place/page'
+
 
 
 
@@ -9,7 +9,6 @@ export default function Home() {
   <>
   <Nav />
   <PostPlace />
-  <SearchPage />
   </>
   )
 }
